Missing Numbers
Numeros, the Artist, had two lists A and B , such that  B was a
permutation of  . Numeros was very proud of these lists. Unfortunately,
while transporting them from one exhibition to another, some numbers
were left out of A . Can you find the missing numbers?

Notes

If a number occurs multiple times in the lists, you must ensure that the
frequency of that number in both lists is the same. If that is not the
case, then it is also a missing number.
You have to print all the missing numbers in ascending order.
Print each missing number once, even if it is missing multiple times.
The difference between maximum and minimum number in B  is less than or
equal to 100 .