Ice Cream Parlour
Sunny and Johnny together have m dollars they want to spend on ice
cream. The parlor offers  flavors, and they want to choose two flavors
so that they end up spending the whole amount.
You are given the cost of these N flavors. The cost of the ith flavor is
denoted by ci. You have to display the indices of the two flavors whose
sum is M .